CleanerModel now calls CleanerService :3
