+++ Famicom/Models/CleanerModel.cs
@@ -9,38 +9,16 @@ namespace Famicom.Models
 {
     public class CleanerModel
     {
-        private readonly ITableControllerService tableControllerService;
-        private readonly TableService tableService;
+        private readonly CleanerService cleanerService;
 
-        public CleanerModel(ITableControllerService tableControllerService, TableService tableService)
+        public CleanerModel(CleanerService cleanerService)
         {
-            this.tableControllerService = tableControllerService;
-            this.tableService = tableService;
+            this.cleanerService = cleanerService;
         }
 
         public async Task UpdateAllTablesMaxHeight(bool isMaxHeight)
         {
-            try
-            {
-                // Retrieve all tables from the database
-                var tables = tableService.GetAllTables();
-
-                if (tables.Count == 0)
-                {
-                    await ReportErrorAsync("No tables found in the database.");
-                    return;
-                }
-            }
-            catch (Exception e)
-            {
-                await ReportErrorAsync(e.Message);
-            }
-        }
-
-        private async Task ReportErrorAsync(string message)
-        {
-            //error reporting logic
-            await Task.CompletedTask;
+            await cleanerService.UpdateAllTablesMaxHeight(isMaxHeight);
         }
     }
 }
